seperated stuff to header files
