+++ src/py_lua.c
@@ -1,85 +1,5 @@
-#include "pylua.h"
+#include "py_lua.h"
 
-
-PyTypeObject pLuaInstance_Type;
-static PyObject* LuaError;
-
-// lua_py.c
-extern lua_State* cL;
-
-
-static PyObject* raise_error(char* msg)
-{
-	if (PyErr_Occurred())
-	{
-		// TODO:
-		// if already error is occured
-		// add a new frame with the following
-		// PyErr_SetString(LuaError, msg);
-		return NULL;
-	}
-
-	PyErr_SetString(LuaError, msg);
-	return NULL;
-}
-
-static PyObject* create_return(lua_State* L, int len)
-{
-	SAVE_STACK_SIZE(L);
-
-	PyObject* pReturn = NULL;
-
-	if (len == 0)
-	{
-		Py_INCREF(Py_None);
-		pReturn = Py_None;
-	}
-	else if (len == 1)
-	{
-		pReturn = PyLua_LuaToPython(L, -1);
-		if (!pReturn)
-		{
-			return raise_error(PERR_CONVERT_LUA_TO_PY);
-		}
-	}
-	else
-	{
-		pReturn = PyTuple_New(len);
-		int tmp = lua_gettop(L);
-
-		for (int i = tmp - len, j = 0; i <= tmp; i++, j++)
-		{
-			PyObject* pItem = PyLua_LuaToPython(L, i);
-			if (!pItem)
-			{
-				return raise_error(PERR_CONVERT_LUA_TO_PY);
-			}
-			PyTuple_SetItem(pReturn, j, pItem);
-		}
-	}
-
-	CHECK_STACK_SIZE(L, 0);
-
-	return pReturn;
-}
-
-static PyObject* flip_boolean(PyObject* ob)
-{
-	if (ob == Py_True)
-	{
-		Py_DECREF(ob);
-		ob = Py_False;
-		Py_INCREF(ob);
-	}
-	else
-	{
-		Py_DECREF(ob);
-		ob = Py_True;
-		Py_INCREF(ob);
-	}
-
-	return ob;
-}
 
 static PyObject* next_LuaCoroutine(PyLua_LuaFunc* self)
 {
